# okito-server/src/main/resources/docs/openapi.json
{
  "components": {
    "schemas": {
      "PostCommentLikeParam": {
        "properties": {
          "commentId": {
            "format": "int64",
            "description": "No comments found.",
            "type": "number",
            "example": "809"
          },
          "userId": {
            "format": "int64",
            "description": "No comments found.",
            "type": "number",
            "example": "216"
          },
          "type": {
            "description": "No comments found.",
            "example": "9579"
          }
        }
      },
...
      "PostLikeParam": {
        "properties": {
          "postId": {
            "format": "int64",
            "description": "No comments found.",
            "type": "number",
            "example": "825"
          },
          "userId": {
            "format": "int64",
            "description": "No comments found.",
            "type": "number",
            "example": "670"
          },
          "type": {
            "description": "No comments found.",
            "example": "12208"
          }
        }
      },
...
      "RespResult": {
        "properties": {
          "timestamp": {
            "format": "int64",
            "description": "No comments found.",
            "type": "number",
            "example": "1677208455247"
          },
          "status": {
            "format": "int32",
            "description": "No comments found.",
            "type": "integer",
            "example": "107"
          },
          "message": {
            "format": "string",
            "description": "No comments found.",
            "type": "string",
            "example": "success"
          },
          "data": {
            "description": "No comments found.(object)",
            "type": "object",
            "example": ""
          }
        }
      },
...
    }
  },
  "servers": [
    {
      "url": "http://localhost:8088/"
    }
  ],
  "openapi": "3.0.3",
  "paths": {
    "/post/comment/like/{userId}": {
      "get": {
        "summary": "select all postCommentLikes by userId",
        "deprecated": false,
        "description": "select all postCommentLikes by userId",
        "tags": [
          "post comment like controller"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespResult"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"timestamp\": 1677208455247,\n  \"status\": 290,\n  \"message\": \"success\",\n  \"data\": {}\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "postcommentlike{userId}",
        "parameters": [
          {
            "schema": {
              "format": "int64",
              "type": "integer"
            },
            "in": "path",
            "description": "userId",
            "required": true,
            "example": "13",
            "name": "userId"
          }
        ]
      }
    },
...
    "/pin/comment/pin/{pinId}": {
      "get": {
        "summary": "select all pinComments by pinId",
        "deprecated": false,
        "description": "select all pinComments by pinId",
        "tags": [
          "pin comment controller"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespResult"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"timestamp\": 1677208455247,\n  \"status\": 783,\n  \"message\": \"success\",\n  \"data\": {}\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "pincommentpin{pinId}",
        "parameters": [
          {
            "schema": {
              "format": "int64",
              "type": "integer"
            },
            "in": "path",
            "description": "pinId",
            "required": true,
            "example": "686",
            "name": "pinId"
          }
        ]
      }
    },
...
    "/post/comment/like": {
      "post": {
        "summary": "add a postCommentLike",
        "deprecated": false,
        "description": "add a postCommentLike",
        "tags": [
          "post comment like controller"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostCommentLikeParam"
              },
              "examples": {
                "json": {
                  "summary": "test data",
                  "value": "{\n  \"commentId\": 447,\n  \"userId\": 266,\n  \"type\": 30929\n}"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespResult"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"timestamp\": 1677208455247,\n  \"status\": 777,\n  \"message\": \"success\",\n  \"data\": {}\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "postcommentlike",
        "parameters": []
      },
      "delete": {
        "summary": "delete postCommentLike by id",
        "deprecated": false,
        "description": "delete postCommentLike by id",
        "tags": [
          "post comment like controller"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespResult"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"timestamp\": 1677208455247,\n  \"status\": 1,\n  \"message\": \"success\",\n  \"data\": {}\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "postcommentlike",
        "parameters": [
          {
            "schema": {
              "format": "int64",
              "type": "integer"
            },
            "in": "query",
            "description": "commentId",
            "required": true,
            "example": "3",
            "name": "commentId"
          },
          {
            "schema": {
              "format": "int64",
              "type": "integer"
            },
            "in": "query",
            "description": "   userId",
            "required": true,
            "example": "629",
            "name": "userId"
          }
        ]
      },
      "put": {
        "summary": "update a postCommentLike",
        "deprecated": false,
        "description": "update a postCommentLike",
        "tags": [
          "post comment like controller"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostCommentLikeParam"
              },
              "examples": {
                "json": {
                  "summary": "test data",
                  "value": "{\n  \"commentId\": 735,\n  \"userId\": 399,\n  \"type\": 24941\n}"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespResult"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"timestamp\": 1677208455247,\n  \"status\": 465,\n  \"message\": \"success\",\n  \"data\": {}\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "postcommentlike",
        "parameters": []
      },
      "get": {
        "summary": "select all postCommentLikes",
        "deprecated": false,
        "description": "select all postCommentLikes",
        "tags": [
          "post comment like controller"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespResult"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"timestamp\": 1677208455247,\n  \"status\": 340,\n  \"message\": \"success\",\n  \"data\": {}\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "postcommentlike",
        "parameters": []
      }
    },
...
    "/post/like": {
      "post": {
        "summary": "add a postLike",
        "deprecated": false,
        "description": "add a postLike",
        "tags": [
          "post like controller"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostLikeParam"
              },
              "examples": {
                "json": {
                  "summary": "test data",
                  "value": "{\n  \"postId\": 168,\n  \"userId\": 40,\n  \"type\": 25113\n}"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespResult"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"timestamp\": 1677208455247,\n  \"status\": 929,\n  \"message\": \"success\",\n  \"data\": {}\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "postlike",
        "parameters": []
      },
      "delete": {
        "summary": "delete postLike by id",
        "deprecated": false,
        "description": "delete postLike by id",
        "tags": [
          "post like controller"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespResult"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"timestamp\": 1677208455247,\n  \"status\": 867,\n  \"message\": \"success\",\n  \"data\": {}\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "postlike",
        "parameters": [
          {
            "schema": {
              "format": "int64",
              "type": "integer"
            },
            "in": "query",
            "description": "postId",
            "required": true,
            "example": "921",
            "name": "postId"
          },
          {
            "schema": {
              "format": "int64",
              "type": "integer"
            },
            "in": "query",
            "description": "userId",
            "required": true,
            "example": "650",
            "name": "userId"
          }
        ]
      },
      "put": {
        "summary": "update postLike",
        "deprecated": false,
        "description": "update postLike",
        "tags": [
          "post like controller"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostLikeParam"
              },
              "examples": {
                "json": {
                  "summary": "test data",
                  "value": "{\n  \"postId\": 407,\n  \"userId\": 115,\n  \"type\": 11588\n}"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespResult"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"timestamp\": 1677208455247,\n  \"status\": 770,\n  \"message\": \"success\",\n  \"data\": {}\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "postlike",
        "parameters": []
      },
      "get": {
        "summary": "select all postLikes",
        "deprecated": false,
        "description": "select all postLikes",
        "tags": [
          "post like controller"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespResult"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"timestamp\": 1677208455247,\n  \"status\": 202,\n  \"message\": \"success\",\n  \"data\": {}\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "postlike",
        "parameters": []
      }
    },
...
    "/post/comment/like/page/{userId}": {
      "get": {
        "summary": "select all postCommentLikes by userId and page",
        "deprecated": false,
        "description": "select all postCommentLikes by userId and page",
        "tags": [
          "post comment like controller"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespResult"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"timestamp\": 1677208455247,\n  \"status\": 275,\n  \"message\": \"success\",\n  \"data\": {}\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "postcommentlikepage{userId}",
        "parameters": [
          {
            "schema": {
              "format": "int64",
              "type": "integer"
            },
            "in": "path",
            "description": "  userId",
            "required": true,
            "example": "89",
            "name": "userId"
          }
        ]
      }
    },
...
    "/pin/comment/count/pin/{pinId}": {
      "get": {
        "summary": "count pinComments by pin id",
        "deprecated": false,
        "description": "count pinComments by pin id",
        "tags": [
          "pin comment controller"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespResult"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"timestamp\": 1677208455247,\n  \"status\": 716,\n  \"message\": \"success\",\n  \"data\": {}\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "pincommentcountpin{pinId}",
        "parameters": [
          {
            "schema": {
              "format": "int64",
              "type": "integer"
            },
            "in": "path",
            "description": "pin id",
            "required": true,
            "example": "217",
            "name": "pinId"
          }
        ]
      }
    },
...
    "/post/like/{userId}": {
      "get": {
        "summary": "select all postLikes by userId",
        "deprecated": false,
        "description": "select all postLikes by userId",
        "tags": [
          "post like controller"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespResult"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"timestamp\": 1677208455247,\n  \"status\": 558,\n  \"message\": \"success\",\n  \"data\": {}\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "postlike{userId}",
        "parameters": [
          {
            "schema": {
              "format": "int64",
              "type": "integer"
            },
            "in": "path",
            "description": "userId",
            "required": true,
            "example": "968",
            "name": "userId"
          }
        ]
      }
    },
...
    "/pin/comment/like/page/{userId}": {
      "get": {
        "summary": "select all pinCommentLikes by userId and page",
        "deprecated": false,
        "description": "select all pinCommentLikes by userId and page",
        "tags": [
          "pin comment like controller"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespResult"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"timestamp\": 1677208455247,\n  \"status\": 469,\n  \"message\": \"success\",\n  \"data\": {}\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "pincommentlikepage{userId}",
        "parameters": [
          {
            "schema": {
              "format": "int64",
              "type": "integer"
            },
            "in": "path",
            "description": "  userId",
            "required": true,
            "example": "458",
            "name": "userId"
          }
        ]
      }
    },
...
    "/pin/like/{userId}": {
      "get": {
        "summary": "select all pinLikes by userId",
        "deprecated": false,
        "description": "select all pinLikes by userId",
        "tags": [
          "pin like controller"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespResult"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"timestamp\": 1677208455247,\n  \"status\": 666,\n  \"message\": \"success\",\n  \"data\": {}\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "pinlike{userId}",
        "parameters": [
          {
            "schema": {
              "format": "int64",
              "type": "integer"
            },
            "in": "path",
            "description": "userId",
            "required": true,
            "example": "917",
            "name": "userId"
          }
        ]
      }
    },
...
    "/pin/like/page/{userId}": {
      "get": {
        "summary": "select all pinLikes by userId and page",
        "deprecated": false,
        "description": "select all pinLikes by userId and page",
        "tags": [
          "pin like controller"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespResult"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"timestamp\": 1677208455247,\n  \"status\": 509,\n  \"message\": \"success\",\n  \"data\": {}\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "pinlikepage{userId}",
        "parameters": [
          {
            "schema": {
              "format": "int64",
              "type": "integer"
            },
            "in": "path",
            "description": "  userId",
            "required": true,
            "example": "462",
            "name": "userId"
          }
        ]
      }
    },
...
    "/pin/like": {
      "post": {
        "summary": "add a pinLike",
        "deprecated": false,
        "description": "add a pinLike",
        "tags": [
          "pin like controller"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PinLikeParam"
              },
              "examples": {
                "json": {
                  "summary": "test data",
                  "value": "{\n  \"pinId\": 848,\n  \"userId\": 264,\n  \"type\": 31122\n}"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespResult"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"timestamp\": 1677208455247,\n  \"status\": 387,\n  \"message\": \"success\",\n  \"data\": {}\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "pinlike",
        "parameters": []
      },
      "delete": {
        "summary": "delete pinLike by id",
        "deprecated": false,
        "description": "delete pinLike by id",
        "tags": [
          "pin like controller"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespResult"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"timestamp\": 1677208455247,\n  \"status\": 612,\n  \"message\": \"success\",\n  \"data\": {}\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "pinlike",
        "parameters": [
          {
            "schema": {
              "format": "int64",
              "type": "integer"
            },
            "in": "query",
            "description": " pinId",
            "required": true,
            "example": "789",
            "name": "pinId"
          },
          {
            "schema": {
              "format": "int64",
              "type": "integer"
            },
            "in": "query",
            "description": "userId",
            "required": true,
            "example": "360",
            "name": "userId"
          }
        ]
      },
      "put": {
        "summary": "update a pinLike",
        "deprecated": false,
        "description": "update a pinLike",
        "tags": [
          "pin like controller"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PinLikeParam"
              },
              "examples": {
                "json": {
                  "summary": "test data",
                  "value": "{\n  \"pinId\": 363,\n  \"userId\": 50,\n  \"type\": 14556\n}"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespResult"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"timestamp\": 1677208455247,\n  \"status\": 429,\n  \"message\": \"success\",\n  \"data\": {}\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "pinlike",
        "parameters": []
      },
      "get": {
        "summary": "select all pinLikes",
        "deprecated": false,
        "description": "select all pinLikes",
        "tags": [
          "pin like controller"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespResult"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"timestamp\": 1677208455247,\n  \"status\": 960,\n  \"message\": \"success\",\n  \"data\": {}\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "pinlike",
        "parameters": []
      }
    },
...
    "/pin/like/single": {
      "get": {
        "summary": "select pinLike by id",
        "deprecated": false,
        "description": "select pinLike by id",
        "tags": [
          "pin like controller"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespResult"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"timestamp\": 1677208455247,\n  \"status\": 771,\n  \"message\": \"success\",\n  \"data\": {}\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "pinlikesingle",
        "parameters": [
          {
            "schema": {
              "format": "int64",
              "type": "integer"
            },
            "in": "query",
            "description": " pinId",
            "required": true,
            "example": "985",
            "name": "pinId"
          },
          {
            "schema": {
              "format": "int64",
              "type": "integer"
            },
            "in": "query",
            "description": "userId",
            "required": true,
            "example": "773",
            "name": "userId"
          }
        ]
      }
    },
...
    "/post/like/single": {
      "get": {
        "summary": "select postLike by id",
        "deprecated": false,
        "description": "select postLike by id",
        "tags": [
          "post like controller"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespResult"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"timestamp\": 1677208455247,\n  \"status\": 972,\n  \"message\": \"success\",\n  \"data\": {}\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "postlikesingle",
        "parameters": [
          {
            "schema": {
              "format": "int64",
              "type": "integer"
            },
            "in": "query",
            "description": "postId",
            "required": true,
            "example": "550",
            "name": "postId"
          },
          {
            "schema": {
              "format": "int64",
              "type": "integer"
            },
            "in": "query",
            "description": "userId",
            "required": true,
            "example": "416",
            "name": "userId"
          }
        ]
      }
    },
    "/post/comment/like/single": {
      "get": {
        "summary": "select postCommentLike by id",
        "deprecated": false,
        "description": "select postCommentLike by id",
        "tags": [
          "post comment like controller"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespResult"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"timestamp\": 1677208455247,\n  \"status\": 626,\n  \"message\": \"success\",\n  \"data\": {}\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "postcommentlikesingle",
        "parameters": [
          {
            "schema": {
              "format": "int64",
              "type": "integer"
            },
            "in": "query",
            "description": "commentId",
            "required": true,
            "example": "874",
            "name": "commentId"
          },
          {
            "schema": {
              "format": "int64",
              "type": "integer"
            },
            "in": "query",
            "description": "   userId",
            "required": true,
            "example": "691",
            "name": "userId"
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "okito",
    "version": "1.0.0"
  },
...
}